rename `marketplace` to `appregistry`
