+++ appregistry/CHANGELOG.md
@@ -1,0 +1,5 @@
+# App Registry App Changelog
+
+## [`v0.1.0`](https://github.com/ignite/apps/releases/tag/appregistry/v0.1.0)
+
+* First release of the app registry app compatible with Ignite >= v28.x.y
