+++ appregistry/README.md
@@ -1,0 +1,36 @@
+# App Registry App
+
+The app registry ignite app lists application from the Ignite App Registry.
+It allows users to discover and install applications built within the Ignite Apps ecosystem.
+
+## Installation
+
+To install the appregistry app, run the following command:
+
+```sh
+ignite app install -g github.com/ignite/apps/appregistry
+```
+
+## List
+
+In order to list the applications, the appregistry app fetches the data from the Ignite App Registry.
+
+```sh
+ignite appregistry list
+```
+
+## Details
+
+To view the details of a specific application, use the following command:
+
+```sh
+ignite appregistry details <app_name>
+```
+
+## Install
+
+To install an application from the appregistry, use the following command:
+
+```sh
+ignite appregistry install <app_name>
+```
