+++ appregistry/cmd/details.go
@@ -9,8 +9,8 @@ import (
 	"github.com/ignite/cli/v28/ignite/pkg/cliui"
 	"github.com/spf13/cobra"
 
-	"github.com/ignite/apps/marketplace/pkg/xgithub"
-	"github.com/ignite/apps/marketplace/registry"
+	"github.com/ignite/apps/appregistry/pkg/xgithub"
+	"github.com/ignite/apps/appregistry/registry"
 )
 
 var (
