+++ appregistry/cmd/list.go
@@ -7,9 +7,9 @@ import (
 	"github.com/ignite/cli/v28/ignite/pkg/cliui"
 	"github.com/spf13/cobra"
 
-	"github.com/ignite/apps/marketplace/pkg/tree"
-	"github.com/ignite/apps/marketplace/pkg/xgithub"
-	"github.com/ignite/apps/marketplace/registry"
+	"github.com/ignite/apps/appregistry/pkg/tree"
+	"github.com/ignite/apps/appregistry/pkg/xgithub"
+	"github.com/ignite/apps/appregistry/registry"
 )
 
 const descriptionLimit = 75
