+++ appregistry/cmd/marketplace.go
@@ -10,18 +10,18 @@ const (
 	githubTokenFlag = "github-token"
 )
 
-// NewMarketplace creates a new marketplace command that holds
-// some other sub commands related to running marketplace like
+// NewAppRegistry creates a new app registry command that holds
+// some other sub commands related to running appregistry like
 // list and info.
-func NewMarketplace() *cobra.Command {
+func NewAppRegistry() *cobra.Command {
 	c := &cobra.Command{
-		Use:     "marketplace [command]",
+		Use:     "appregistry [command]",
 		Aliases: []string{"mp"},
-		Short:   "Browse the Ignite App Marketplace",
-		Long: `Marketplace is a command line tool that helps you to search for ignite apps.
+		Short:   "Browse the Ignite App Registry App",
+		Long: `AppRegistry is a command line tool that helps you to search for ignite apps.
 It also helps you to get more information about an app.
 Please note this command uses the Github API that a very limited rate limit for unauthenticated requests
-so it's recommended to use the --github-token flag you want to use marketplace commands frequently.`,
+so it's recommended to use the --github-token flag you want to use appregistry commands frequently.`,
 		SilenceErrors: true,
 	}
