+++ appregistry/registry/details.go
