+++ appregistry/registry/types.go
