--- marketplace/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Marketplace App Changelog
-
-## [`v0.1.0`](https://github.com/ignite/apps/releases/tag/marketplace/v0.1.0)
-
-* First release of the Markplace app compatible with Ignite >= v28.x.y
