--- marketplace/README.md
@@ -1,36 +1,0 @@
-# Marketplace App
-
-The marketplace ignite app lists application from the Ignite App Registry.
-It allows users to discover and install applications built within the Ignite Apps ecosystem.
-
-## Installation
-
-To install the marketplace app, run the following command:
-
-```sh
-ignite app install -g github.com/ignite/apps/marketplace
-```
-
-## List
-
-In order to list the applications, the marketplace app fetches the data from the Ignite App Registry.
-
-```sh
-ignite marketplace list
-```
-
-## Details
-
-To view the details of a specific application, use the following command:
-
-```sh
-ignite marketplace details <app_name>
-```
-
-## Install
-
-To install an application from the marketplace, use the following command:
-
-```sh
-ignite marketplace install <app_name>
-```
